Refactor validation check and remove default password rules
This commit refactors the function that checks validation rules. Instead of checking if the `$validation_rules` variable is `null`, it now checks if the variable is `empty`. This allows the function to proceed correctly even with non-null yet empty variables.

Moreover, the creation of the validator was simplified into a single line. This update makes the code more readable and concise.

In SimpleSecretsServiceProvider, the default password validation rules were removed. This renders more flexibility, allowing projects using the simple-secrets package to set their own password validation rules as per their requirements.